refactor(exception): move package
