+++ src/main/java/lotto/exception/purchaseamountexception/MaxPurchaseExceedException.java
@@ -1,4 +1,7 @@
-package lotto.exception;
+package lotto.exception.purchaseamountexception;
+
+import lotto.exception.ErrorConstants;
+import lotto.exception.ErrorMessage;
 
 public class MaxPurchaseExceedException extends IllegalStateException {
 	public MaxPurchaseExceedException() {
