+++ src/main/java/lotto/exception/purchaseamountexception/NotDivisibleByLottoPriceException.java
@@ -1,4 +1,7 @@
-package lotto.exception;
+package lotto.exception.purchaseamountexception;
+
+import lotto.exception.ErrorConstants;
+import lotto.exception.ErrorMessage;
 
 public class NotDivisibleByLottoPriceException extends IllegalArgumentException {
 	public NotDivisibleByLottoPriceException() {
